+++ packages/eth-json-rpc-provider/src/safe-event-emitter-provider.ts
@@ -1,6 +1,49 @@
 import type { JsonRpcEngine } from '@metamask/json-rpc-engine';
+import { JsonRpcError } from '@metamask/rpc-errors';
 import SafeEventEmitter from '@metamask/safe-event-emitter';
-import type { JsonRpcRequest } from '@metamask/utils';
+import type {
+  Json,
+  JsonRpcId,
+  JsonRpcParams,
+  JsonRpcRequest,
+  JsonRpcVersion2,
+} from '@metamask/utils';
+import { v4 as uuidV4 } from 'uuid';
+
+/**
+ * A JSON-RPC request conforming to the EIP-1193 specification.
+ */
+type Eip1193Request<Params extends JsonRpcParams> = {
+  id?: JsonRpcId;
+  jsonrpc?: JsonRpcVersion2;
+  method: string;
+  params?: Params;
+};
+
+/**
+ * Converts an EIP-1193 request to a JSON-RPC request.
+ *
+ * @param eip1193Request - The EIP-1193 request to convert.
+ * @returns The corresponding JSON-RPC request.
+ */
+export function convertEip1193RequestToJsonRpcRequest<
+  Params extends JsonRpcParams,
+>(
+  eip1193Request: Eip1193Request<Params>,
+): JsonRpcRequest<Params | Record<never, never>> {
+  const {
+    id = uuidV4(),
+    jsonrpc = '2.0',
+    method,
+    params = {},
+  } = eip1193Request;
+  return {
+    id,
+    jsonrpc,
+    method,
+    params,
+  };
+}
 
 /**
  * An Ethereum provider.
@@ -31,30 +74,67 @@ export class SafeEventEmitterProvider extends SafeEventEmitter {
   /**
    * Send a provider request asynchronously.
    *
-   * @param req - The request to send.
-   * @param callback - A function that is called upon the success or failure of the request.
+   * @param eip1193Request - The request to send.
+   * @returns The JSON-RPC response.
    */
-  sendAsync = (
-    req: JsonRpcRequest,
+  async request<Params extends JsonRpcParams, Result extends Json>(
+    eip1193Request: Eip1193Request<Params>,
+  ): Promise<Result> {
+    const jsonRpcRequest =
+      convertEip1193RequestToJsonRpcRequest(eip1193Request);
+    const response = await this.#engine.handle<
+      Params | Record<never, never>,
+      Result
+    >(jsonRpcRequest);
+
+    if ('result' in response) {
+      return response.result;
+    }
+
+    const error = new JsonRpcError(
+      response.error.code,
+      response.error.message,
+      response.error.data,
+    );
+    if ('stack' in response.error) {
+      error.stack = response.error.stack;
+    }
+    throw error;
+  }
+
+  /**
+   * Send a provider request asynchronously.
+   *
+   * This method serves the same purpose as `request`. It only exists for
+   * legacy reasons.
+   *
+   * @param eip1193Request - The request to send.
+   * @param callback - A function that is called upon the success or failure of the request.
+   * @deprecated Please use `request` instead.
+   */
+  sendAsync = <Params extends JsonRpcParams>(
+    eip1193Request: Eip1193Request<Params>,
     // TODO: Replace `any` with type
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     callback: (error: unknown, providerRes?: any) => void,
   ) => {
-    this.#engine.handle(req, callback);
+    const jsonRpcRequest =
+      convertEip1193RequestToJsonRpcRequest(eip1193Request);
+    this.#engine.handle(jsonRpcRequest, callback);
   };
 
   /**
    * Send a provider request asynchronously.
    *
-   * This method serves the same purpose as `sendAsync`. It only exists for
+   * This method serves the same purpose as `request`. It only exists for
    * legacy reasons.
    *
-   * @deprecated Use `sendAsync` instead.
-   * @param req - The request to send.
+   * @param eip1193Request - The request to send.
    * @param callback - A function that is called upon the success or failure of the request.
+   * @deprecated Please use `request` instead.
    */
-  send = (
-    req: JsonRpcRequest,
+  send = <Params extends JsonRpcParams>(
+    eip1193Request: Eip1193Request<Params>,
     // TODO: Replace `any` with type
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     callback: (error: unknown, providerRes?: any) => void,
@@ -62,6 +142,8 @@ export class SafeEventEmitterProvider extends SafeEventEmitter {
     if (typeof callback !== 'function') {
       throw new Error('Must provide callback to "send" method.');
     }
-    this.#engine.handle(req, callback);
+    const jsonRpcRequest =
+      convertEip1193RequestToJsonRpcRequest(eip1193Request);
+    this.#engine.handle(jsonRpcRequest, callback);
   };
 }